+++ src/core/prelude/object/iterators/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.35.0 (2021-03-04)
+
+#### :rocket: New Feature
+
+* Added a new option `withNonEnumerables` `Object.forEach`
+
 ## v3.20.0 (2020-07-05)
 
 #### :rocket: New Feature