+++ src/core/prelude/object/mixin/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.35.0 (2021-03-04)
+
+#### :rocket: New Feature
+
+* Added a new option `withNonEnumerables` `Object.mixin`
+
 ## v3.34.0 (2021-03-03)
 
 #### :boom: Breaking Change