+++ src/core/request/engines/provider/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.35.0 (2021-03-04)
+
+#### :bug: Bug Fix
+
+* Fixed incorrect path resolving by a chain of provider engines with parameters within URL
+
 ## v3.33.0 (2021-02-26)
 
 #### :rocket: New Feature